% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.phnorm.r
\name{mice.impute.phnorm}
\alias{mice.impute.phnorm}
\alias{phnorm}
\title{Imputation by GLM and phylogenetic information}
\usage{
mice.impute.phnorm(y, ry, x, psi, psiinv, ...)
}
\arguments{
\item{y}{Incomplete data vector of length \code{n}}

\item{ry}{Vector of missing data pattern (\code{FALSE}=missing,
\code{TRUE}=observed)}

\item{x}{Matrix (\code{n} x \code{p}) of complete covariates.}

\item{psi}{Covariance matrix containing phylogenetic information.}

\item{psiinv}{Inverse of \code{psi}.}

\item{...}{Other named arguments.}
}
\value{
A vector of length \code{nmis} with imputations.
}
\description{
Imputes univariate continuous missing data using the generalized least square approach.
}
\note{
\code{mice.impute.phnorm} is based on the
\code{mice.impute.norm} method from \code{mice} package and the GLM
approach by Garland and Ives (2000).
}
\author{
Patrik Drhlik, Simon P. Blomberg, 2016
}
\references{
Van Buuren, S., Groothuis-Oudshoorn, K. (2011). \code{mice}:
Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of
Statistical Software}, \bold{45}(3), 1-67.
\url{http://www.jstatsoft.org/v45/i03/}

Garland Jr, Theodore, and Anthony R. Ives. Using the past to predict
the present: confidence intervals for regression equations in phylogenetic
comparative methods. \emph{The American Naturalist}, 155.3 (2000): 346-364.
\url{http://www.jstor.org/stable/10.1086/303327}
}

