% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomice.R
\docType{package}
\name{phylomice}
\alias{phylomice}
\alias{phylomice-package}
\title{phylomice: Multiple imputation using phylogenetic information}
\description{
The phylomice package provides methods for imputing missing
data using phylogenetic information. Currently contained
methods:
}
\details{
\bold{phnorm}
\bold{phpmm} - under development
}

