% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomice.R
\name{precomputePsi}
\alias{precomputePsi}
\title{Precomputes psi and psiinv given a tree.}
\usage{
precomputePsi(tree)
}
\arguments{
\item{tree}{Phylogenetic tree that can be processed
by the ape library.}
}
\value{
Returns a list with components psi
and psiinv that are needed in phylogenetic
imputation methods.
}
\description{
Precomputes psi and psiinv given a tree.
}

